+++ src/components/Contact/ContactList/ContactList.js
@@ -2,14 +2,13 @@ import React from 'react';
 import './ContactList.scss';
 import NavBar from '../../Navbar';
 import { NavLink } from 'react-router-dom';
-import EnhancedTable from './components/table';
+import EnhancedTable from './components/Table';
 
 const ContactList = () => (
   // TODO1: 需要一个弹窗输入信息
   // TODO2: 需要一个Navbar切换表格
-  // TODO3: 需要一个按钮读入CSV文件
-  // TODO4: 模块化代码
-  // TODO5: 完成后端数据库
+  // TODO4: 完成后端数据库
+  // TODO5: 需要实现联系人等属性的跳转
   // TODO6: 将添加、删除用数据库实现
   <div>
     <header>
