+++ src/components/Contact/ContactList/components/MuiTheme.js
@@ -1,0 +1,17 @@
+import { createMuiTheme } from '@material-ui/core/styles';
+
+
+// 表格部分样式
+const Theme = createMuiTheme({
+    palette: {
+      primary: {
+        main: '#4caf50',
+      },
+      secondary: {
+        main: '#ff9100',
+      },
+    },
+  
+});
+
+export default Theme;
