--- src/components/Contact/ContactList/components/enhancedTableToolbar.js
@@ -1,80 +1,0 @@
-import React from 'react';
-
-import { lighten, makeStyles } from '@material-ui/core/styles';
-import Toolbar from '@material-ui/core/Toolbar';
-import Typography from '@material-ui/core/Typography';
-import IconButton from '@material-ui/core/IconButton';
-import Tooltip from '@material-ui/core/Tooltip';
-import DeleteIcon from '@material-ui/icons/Delete';
-import FilterListIcon from '@material-ui/icons/FilterList';
-import PropTypes from 'prop-types';
-import clsx from 'clsx';
-
-
-const useToolbarStyles = makeStyles((theme) => ({
-    root: {
-      paddingLeft: theme.spacing(2),
-      paddingRight: theme.spacing(1),
-    },
-    highlight:
-      theme.palette.type === 'light'
-        ? {
-            color: theme.palette.secondary.main,
-            // backgroundColor: lighten(theme.palette.secondary.light, 0.85),
-            backgroundColor: "#A8D8B9"
-          }
-        : {
-            color: theme.palette.text.primary,
-            // backgroundColor: theme.palette.secondary.dark,
-            backgroundColor: "#A8D8B9"
-          },
-    title: {
-      flex: '1 1 100%',
-    },
-  }));
-  
-const EnhancedTableToolbar = (props) => {
-    const classes = useToolbarStyles();
-    const { numSelected, rowCount } = props;
-  
-    return (
-      <Toolbar
-        className={clsx(classes.root, {
-          [classes.highlight]: numSelected > 0,
-        })}
-      >
-        {numSelected > 0 ? (
-          <Typography className={classes.title} color="textSecondary" variant="subtitle1" component="div">
-            {numSelected} selected
-          </Typography>
-        ) : (
-          <Typography className={classes.title} variant="h4" id="tableTitle" component="div">
-            Contacts
-            <Typography className={classes.title} align="left" variant="caption" id="rowNum" component="div">
-                {rowCount} contacts
-            </Typography>
-          </Typography>
-        )}
-  
-        {numSelected > 0 ? (
-          <Tooltip title="Delete">
-            <IconButton aria-label="delete">
-              <DeleteIcon />
-            </IconButton>
-          </Tooltip>
-        ) : (
-          <Tooltip title="Filter list">
-            <IconButton aria-label="filter list">
-              <FilterListIcon />
-            </IconButton>
-          </Tooltip>
-        )}
-      </Toolbar>
-    );
-  };
-  
-  EnhancedTableToolbar.propTypes = {
-    numSelected: PropTypes.number.isRequired,
-  };
-
-  export default EnhancedTableToolbar;
