+++ src/components/Contact/ContactList/services/getDate.js
@@ -1,0 +1,12 @@
+import React, { Component } from 'react';
+
+function getDate() {
+    let today = new Date();
+    let dd = String(today.getDate()).padStart(2, '0');
+    let mm = String(today.getMonth() + 1).padStart(2, '0'); //January is 0!
+    let yyyy = today.getFullYear();
+    today = mm + '/' + dd + '/' + yyyy;
+    return today;
+}
+
+export default getDate;
