+++ src/components/Contact/ContactList/services/updateRow.js
@@ -1,0 +1,10 @@
+import React, { Component } from 'react';
+
+const updateRow = (newData) => {
+    if (newData.contactOwner === '') {
+        newData.contactOwner = 'Unassigned';
+    }
+    return newData;
+}
+
+export default updateRow;
